MS: removed back compat for 1.5.2, 1.6.2; started deprecation updates

+++ trunk/MagicSpells/src/com/nisovin/magicspells/spells/targeted/CombustSpell.java
@@ -59,7 +59,7 @@ public class CombustSpell extends TargetedSpell implements TargetedEntitySpell {
 	private boolean combust(Player player, final LivingEntity target, float power) {
 		if (target instanceof Player && checkPlugins && player != null) {
 			// call other plugins
-			EntityDamageByEntityEvent event = new EntityDamageByEntityEvent(player, target, DamageCause.ENTITY_ATTACK, 1);
+			EntityDamageByEntityEvent event = new EntityDamageByEntityEvent(player, target, DamageCause.ENTITY_ATTACK, (double)1);
 			Bukkit.getServer().getPluginManager().callEvent(event);
 			if (event.isCancelled()) {
 				return false;
